Added deletion of license function, text sanitization

+++ legalzard_admin_temp/static/assets/js/licenses.js
@@ -2,6 +2,7 @@ let index = 0;
 let updateIndex = 0;
 let fileData = {};
 let licenseTagAddCount = 0;
+let responseStatus = 400;
 
 document.addEventListener("DOMContentLoaded", function(event){
 
@@ -37,44 +38,49 @@ document.addEventListener("DOMContentLoaded", function(event){
 
 
     // BEGIN delete of license detail
-    document.querySelector("#confirm-delete").addEventListener("click", function(event){
+    deleteLicenseConfirmEl = document.querySelector("#confirm-delete");
+    if (deleteLicenseConfirmEl) {
 
-        const spinner = `<span class="spinner-border spinner-border-sm" role="status" aria-hidden="true"></span>
-                    Deleting...`
+        deleteLicenseConfirmEl.addEventListener("click", function(event){
 
-
-        const eventId = document.getElementById("confirm-delete").getAttribute("data-id");
-        const licenseName = document.getElementById("confirm-delete").getAttribute("data-license-name");
-        this.innerHTML = spinner;
-        const errorEl = document.getElementById("delete-license-error-notify");
-        errorEl.innerHTML = "";
-
-        fetch(`/api/licenses/${eventId}/`, {
-                method: "DELETE",
-                headers: {"Content-Type": "application/json"}
-            }).then(response => {
-
-                if (response.status === 200){
-                    const licenseDetailContentEl = document.getElementById(`license-table-row-${eventId}`)
-                    licenseDetailContentEl.remove();
-                    this.innerHTML = "Delete";
-                    document.getElementById("delete-license-modal-close").click();
-
-                }else {
-
-                    this.innerHTML="Delete";
-                    response.json()
-                }
-
-            }).then(jsonData => {
-
-                errorEl.innerHTML = jsonData.error_msg;
-
-            }).catch(error => {
-
+            const spinner = `<span class="spinner-border spinner-border-sm" role="status" aria-hidden="true"></span>
+                        Deleting...`
+    
+    
+            const eventId = document.getElementById("confirm-delete").getAttribute("data-id");
+            const licenseName = document.getElementById("confirm-delete").getAttribute("data-license-name");
+            this.innerHTML = spinner;
+            const errorEl = document.getElementById("delete-license-error-notify");
+            errorEl.innerHTML = "";
+    
+            fetch(`/api/licenses/${eventId}/`, {
+                    method: "DELETE",
+                    headers: {"Content-Type": "application/json"}
+                }).then(response => {
+    
+                    if (response.status === 200){
+                        const licenseDetailContentEl = document.getElementById(`license-table-row-${eventId}`)
+                        licenseDetailContentEl.remove();
+                        this.innerHTML = "Delete";
+                        document.getElementById("delete-license-modal-close").click();
+    
+                    }else {
+    
+                        this.innerHTML="Delete";
+                        response.json()
+                    }
+    
+                }).then(jsonData => {
+    
+                    errorEl.innerHTML = jsonData.error_msg;
+    
+                }).catch(error => {
+    
+                })
             })
-        })
-        // END delete of license detail
+            // END delete of license detail
+
+    }
   
 
 
@@ -204,8 +210,8 @@ const saveDataToDatabase = (event) =>{
             if (otherLicenseAttribute.includes(",")){
 
                 for (let attribute of otherLicenseAttribute.split(",")){
-                    licenseAttributeList.push(attribute.trim());
-                    otherLicenseAttributeList.push(attribute.trim());
+                    licenseAttributeList.push(sanitizeText(attribute.trim()));
+                    otherLicenseAttributeList.push(sanitizeText(attribute.trim()));
                 }
             }else{
                 licenseAttributeList.push(otherLicenseAttribute);
@@ -224,11 +230,11 @@ const saveDataToDatabase = (event) =>{
             license_url: licenseUrl,
             type_of_license: typeOfLicense,
             license_tags: getLicenseTagContent(),
-            short_description: shortDescription,
-            description: licenseDescription,
-            disclaimer: disclaimer,
-            risk_for_choosing_license: riskForChoosingLicense,
-            limitation_of_liability: limitationOfLiability,
+            short_description: sanitizeText(shortDescription),
+            description: sanitizeText(licenseDescription),
+            disclaimer: sanitizeText(disclaimer),
+            risk_for_choosing_license: sanitizeText(riskForChoosingLicense),
+            limitation_of_liability: sanitizeText(limitationOfLiability),
             logo_detail: {
                 filename: fileData.filename,
                 actual_filename: fileData.actual_filename,
@@ -237,7 +243,7 @@ const saveDataToDatabase = (event) =>{
             },
             recommendation: " ",
             license_attributes: {
-                heading: licenseAttributeHeading,
+                heading: sanitizeText(licenseAttributeHeading),
                 attributes: licenseAttributeList
             },
             license_compatible_with_lookup: licenseCompatibleWithList,
@@ -259,29 +265,34 @@ const saveDataToDatabase = (event) =>{
             body: JSON.stringify(data),
             headers: {"Content-Type": "application/json"}
         }).then(function(response){
+
+            responseStatus = response.status;
             if (response.status === 201 || response.status === 200){
 
                 window.location.href = "/temp-admin/licenses/";
 
             }else{
 
+                return response.json();
+
+            }
+
+        }).then(jsonData => {
 
                 // set error 
                 errorContainerEl.style.display = "block";
-                errorContentEl.textContent = "Something went wrong, whilst saving the data!";
+                errorContentEl.textContent = jsonData.error_msg;
 
                 // deactivate loading
                 btnSaveData.innerHTML = "Save";
                 btnSaveData.disabled = false;
 
-
-            }
-
         }).catch(function(err){
 
             // set error 
+            console.log(err)
             errorContainerEl.style.display = "block";
-            errorContentEl.textContent = "Something went wrong, check your network!";
+            errorContentEl.textContent = err.toString();
 
             // deactivate loading
             btnSaveData.innerHTML = "Save";
@@ -725,5 +736,9 @@ const deleteLicense = () => {
 
     });
   }
+
+const sanitizeText = (text) => {
+    return text.replace(/["${}]/g,"");
+}
